Gameboard and Cell factory functions
a gameboard function with the layout of the 3x3
board using for and nested for loops to create an
array with 3 nested arrays representing the rows
and 3 Cell() within the row arrays for the column.
The Cell factory functions stores the value of
each cell starting at 0 for a brand new game and
the addToken function to alter cell value to
show player input and a getValue function to check
cell value.