markCell and printBoard function in Gameboard
markCell function checks if the cell in the
selected row and column is already marked and
marks it with the player's value if it's free.
printBoard function logs the status of the board
in the console for playing the game before ui is
built. (can delete later). Gameboard factory now
has a getBoard, markCell, and printBoard function
stored.