checkWin function using for loops
- for loop checks each row's(i) cell.getValue()
to see if it matches the current player's token

- for loop checks each column's(j) cell.getValue()
to see if it matches current player's token

- two if statements for specific diagonal cells'
values to see if it matches current player

- set fullBoard to true and check every cell for
a getValue() of 0 indicating unmarked cell. If
there is an unmarked cell fullBoard = false. If
no getValue() of 0 board is full and it's a 'Tie'